ring: Keep the image-resident mirror out of the hot scalars
The mirror was declared beside m_start/m_end, on the reasoning that the
print path wants it in a cache line it already has. Measuring both
placements says the opposite: sitting there displaces m_writable, m_mask
and m_array, and costs the BULK writer - the run-at-a-time path that
handles ordinary ASCII - about 8 percent, against a 3 to 6 percent gain on
the per-character path it was meant to help. A build with the fields moved
back measured at parity, and a build padding the same offset with dead
bytes reproduced most of the loss, so it is the layout and not the flag.

Move them back among the image fields. Ordinary text is the commoner
workload and it should not pay for the double-width path.

Also compare the image bottom in signed long in both drop helpers.
get_bottom() is int while row_t is unsigned long, so the int was promoted
and a negative bottom would have compared as a very large positive.

+++ src/ring.cc
@@ -302,7 +302,7 @@ Ring::drop_images_before(row_t row) noexcept
         for (auto it = m_image_by_top_map.begin();
              it != m_image_by_top_map.end() && it->first < row; ) {
                 auto const image = it->second;
-                if (image->get_bottom() >= row) {
+                if (long(image->get_bottom()) >= long(row)) {
                         ++it;
                         continue;
                 }
